requirements.txt
-i https://pypi.org/simple
appdirs==1.4.4
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
idna==3.4; python_version >= '3.5'
importlib-metadata==6.8.0; python_version >= '3.8'
pyee==8.2.2
pyppeteer==1.0.2
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1; python_version >= '3.7'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
websockets==10.4; python_version >= '3.7'
zipp==3.16.2; python_version >= '3.8'
